% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_travis_user}
\alias{add_travis_user}
\title{Add a user via their Travis repo.}
\usage{
add_travis_user(travis_repo, email, vault = NULL)
}
\arguments{
\item{travis_repo}{Name of Travis repository, usually in a format
\code{<<username>>/<<repo>>}}

\item{email}{Email address of the user. This is used to identify
users.}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
On Travis, every repo has a private/public key pair. This function adds a
user and downloads the public key from Travis.
}
\examples{
vault <- file.path(tempdir(), ".vault")
create_vault(vault)

add_travis_user("gaborcsardi/secret", vault = vault)
list_users(vault = vault)
delete_user("travis-gaborcsardi-secret", vault = vault)
}
\seealso{
Other user functions: \code{\link{add_github_user}},
  \code{\link{add_user}}, \code{\link{delete_user}},
  \code{\link{list_users}}
}
