% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user}
\alias{add_user}
\title{Add a new user to the vault}
\usage{
add_user(email, public_key, vault = NULL)
}
\arguments{
\item{email}{Email address of the user. This is used to identify
users.}

\item{public_key}{Public key of the user. This is used to encrypt
the secrets for the different users. It can be
\itemize{
\item a string containing a PEM,
\item a file name that points to a PEM file,
\item a \code{pubkey} object created via the \code{openssl} package.
}}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
By default the new user does not have access to any secrets.
See \code{\link[=add_secret]{add_secret()}} or \code{\link[=share_secret]{share_secret()}} to give them access.
}
\seealso{
Other user functions: \code{\link{add_github_user}},
  \code{\link{add_travis_user}}, \code{\link{delete_user}},
  \code{\link{list_users}}
}
