% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{create_package_vault}
\alias{create_package_vault}
\title{Create a vault in an R package}
\usage{
create_package_vault(path = ".")
}
\arguments{
\item{path}{Path to the R package. A file or directory within the
package is fine, too.}
}
\value{
The directory of the vault, invisibly.
}
\description{
Create a vault in an R package
}
