% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret}
\alias{delete_secret}
\title{Remove a secret from the vault}
\usage{
delete_secret(name, vault = NULL)
}
\arguments{
\item{name}{Name of the secret to delete.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Remove a secret from the vault
}
