% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user}
\usage{
delete_user(email, vault = NULL)
}
\arguments{
\item{email}{Email address of the user.}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
It also removes access of the user to all secrets, so if the user
is re-added again, they will not have access to any secrets.
}
\seealso{
Other user functions: \code{\link{add_github_user}},
  \code{\link{add_travis_user}}, \code{\link{add_user}},
  \code{\link{list_users}}
}
