% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{eval_prop}
\alias{eval_prop}
\title{cell type proportion results evaluation}
\usage{
eval_prop(trueP, estP)
}
\arguments{
\item{trueP}{true proportion}

\item{estP}{estimated proportion}
}
\value{
RMSD, mAD, R
}
\description{
cell type proportion results evaluation
}
