% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_github_key}
\alias{get_github_key}
\title{Get the SSH public key of a GitHub user}
\usage{
get_github_key(github_user, i = 1)
}
\arguments{
\item{github_user}{GitHub username.}

\item{i}{Which key to get, in case the user has multiple keys.
\code{get_github_key()} retrieves the first key by default.}
}
\value{
Character scalar.
}
\description{
Get the SSH public key of a GitHub user
}
