% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{get_secret}
\alias{get_secret}
\title{Retrieve a secret from the vault}
\usage{
get_secret(key, vault = NULL)
}
\arguments{
\item{key}{Name of the secret.}

\item{vault}{Vault location. TODO}
}
\description{
Retrieve a secret from the vault
}
