% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_travis_key}
\alias{get_travis_key}
\title{Retrieve the public key of a Travis CI repository}
\usage{
get_travis_key(travis_repo)
}
\arguments{
\item{travis_repo}{The repository slug, e.g. \code{gaborcsardi/secret}.}
}
\value{
Character scalar, the key. If the repository does not exist,
or it is not user in Travis CI, an HTTP 404 error is thrown.
}
\description{
Retrieve the public key of a Travis CI repository
}
