% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{get_user_file}
\alias{get_user_file}
\title{Get the file of a user (email)}
\usage{
get_user_file(vault, email)
}
\arguments{
\item{vault}{Vault directory.}

\item{email}{Email address (or user name, in general).}
}
\value{
The path to the user's public key. (It might not exist yet.)
}
\description{
We assume that \code{vault} is a proper vault directory, and \code{email} is a
valid email address.
}
\keyword{internal}
