% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{local_key}
\alias{local_key}
\title{Reads local secret key.}
\usage{
local_key()
}
\description{
Reads a local secret key from disk. The location of this file is specified in the environment variable called \code{USER_KEY} (See details).  If this environment variable does not exist, then attempts to read the key at:
\itemize{
\item \code{~/.ssh/id_rsa}
\item \code{~/.ssh/id_rsa.pem}.
}
}
\details{
The location of the key is defined by:

\code{Sys.getenv("USER_KEY")}

To use a local in a different location, set an environment variable:

\code{Sys.setenv(USER_KEY = "path/to/private/key")}
}
\seealso{
Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{get_secret}},
  \code{\link{list_secrets}}, \code{\link{share_secret}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
