% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SECRET.R
\name{normcount}
\alias{normcount}
\title{Normalization}
\usage{
normcount(count, method)
}
\arguments{
\item{count}{the dataset need to be normalized}

\item{method}{select which normalization method to use}
}
\value{
normalized dataset
}
\description{
Normalization
}
