% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ secret }
\alias{ secret }
\alias{ secret }
\title{ Share Sensitive Information in R Packages }
\description{
Allow sharing sensitive information, for example passwords, 'API' keys, etc.,
  in R packages, using public key cryptography.
}
