% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{scRefer}
\alias{scRefer}
\title{generate cell type reference}
\usage{
scRefer(input_data, ct_var, sample_var)
}
\arguments{
\item{input_data}{input single cell data}

\item{ct_var}{what variable name for cell type}

\item{sample_var}{what variable name for sample}
}
\value{
gene by cell type expression
}
\description{
generate cell type reference
}
