% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\docType{package}
\name{secret-package}
\alias{secret-package}
\title{The secret package}
\description{
The secret package
}
\details{
Hide the source code of functions in package via Rcpp.
}
\seealso{
\link{https://stackoverflow.com/a/1360175/2906900}
}
