% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SECRET.R
\name{SECRET}
\alias{SECRET}
\title{Semi-reference based cell type deconvolution method}
\usage{
SECRET(
  bulk,
  reference,
  withUnknown = T,
  w = w,
  yNorm,
  bNorm,
  exprs_mode = F,
  bulkfull = NULL
)
}
\arguments{
\item{bulk}{raw filtered bulk data}

\item{reference}{raw filtered cell type reference data}

\item{withUnknown}{whether include unknown cell type}

\item{w}{weight for each genes}

\item{yNorm}{normalization method}

\item{bNorm}{normalization method}

\item{exprs_mode}{whether to calculate gene expression, this step takes long time so the default is F}

\item{bulkfull}{raw full bulk data}
}
\value{
estimated cell type proportion or cell type gene expression if exprs_mode=T
}
\description{
Semi-reference based cell type deconvolution method
}
