% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{share_secret}
\alias{share_secret}
\alias{everyone}
\alias{noone}
\title{Share a secret among some users}
\usage{
share_secret(name, users = everyone(), vault = NULL)

everyone()

noone()
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{users}{Character vector of usernames or email addresses
to share the secret with. The special values \code{\link[=everyone]{everyone()}} and
\code{\link[=noone]{noone()}} make it easy to share a secret with all users, or no
users at all.}

\item{vault}{Vault location. TODO}
}
\description{
Use this function to restrict or extend the set of users that
have access to a secret. The calling user must have access to
the secret as well.
}
