% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{share_secret_with_key}
\alias{share_secret_with_key}
\title{Share a secret, its AES key is known already.}
\usage{
share_secret_with_key(name, users, aeskey, vault)
}
\arguments{
\item{name}{Name of the secret.}

\item{users}{Email addresses of users.}

\item{aeskey}{AES key of the secret.}

\item{vault}{Vault directory.}
}
\description{
Share a secret, its AES key is known already.
}
\keyword{internal}
