% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{share_secret_with_key1}
\alias{share_secret_with_key1}
\title{Share a secret with a single user, AES key is known.}
\usage{
share_secret_with_key1(name, email, aeskey, vault)
}
\arguments{
\item{name}{Name of the secret.}

\item{email}{Email address of the user.}

\item{aeskey}{The AES key of the secret.}

\item{vault}{Vault directory.}
}
\description{
Share a secret with a single user, AES key is known.
}
\keyword{internal}
