% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{store_secret_with_key}
\alias{store_secret_with_key}
\title{Store a secret, encrypted with its AES key}
\usage{
store_secret_with_key(name, value, key, vault)
}
\arguments{
\item{name}{Name of secret.}

\item{value}{Value of secret.}

\item{key}{The AES key, an \code{aes} object from the \code{openssl} package.}

\item{vault}{Vault directory.}
}
\description{
Store a secret, encrypted with its AES key
}
\keyword{internal}
