% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{try_get_aes_key}
\alias{try_get_aes_key}
\title{Try to get the AES key of a secret, using a private RSA key}
\usage{
try_get_aes_key(vault, name, key)
}
\description{
We just try the private key against all encrypted copies of the
AES key. If none of the succeed, then we return \code{NULL}. Otherwise
we return the AES key, an \code{aes} object, from the \code{openssl} package.
}
\keyword{internal}
