% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{unshare_secret}
\alias{unshare_secret}
\title{Unshare a secret among some users}
\usage{
unshare_secret(name, users, vault = NULL)
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{users}{Character vector of usernames or email addresses
to unshare the secret with.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Use this function to restrict the set of users that have access to a
secret. Note that users may still have access to the secret, through
version control history, or if they have a copy of the project. They
will not have access to future values of the secret, though.
}
\seealso{
\code{\link[=share_secret]{share_secret()}}

Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{get_secret}},
  \code{\link{list_secrets}}, \code{\link{share_secret}},
  \code{\link{update_secret}}
}
