% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{update_secret}
\alias{update_secret}
\title{Update a secret in the vault.}
\usage{
update_secret(key, vault = NULL)
}
\arguments{
\item{key}{Name of the secret.}

\item{vault}{Vault location. TODO}
}
\description{
Update a secret in the vault.
}
