% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_github_user}
\alias{add_github_user}
\title{Add a user via their GitHub username}
\usage{
add_github_user(github_user, email = NULL, vault = NULL, i = 1)
}
\arguments{
\item{github_user}{User name on GitHub.}

\item{email}{Email address. If not supplied, constructs an email address
with the GitHub user name: \code{github-username}}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}

\item{i}{Integer, indicating which GitHub key to use (if more than one
GitHub key exists).}
}
\description{
Add a user via their GitHub username
}
\seealso{
Other user functions: \code{\link{add_travis_user}},
  \code{\link{add_user}}, \code{\link{delete_user}},
  \code{\link{list_users}}
}
