% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_github_user}
\alias{add_github_user}
\title{Add a user via their GitHub username}
\usage{
add_github_user(github_id, email = NULL, vault = NULL)
}
\arguments{
\item{github_id}{User id on GitHub.}

\item{email}{Email address. This can be usually queried from GitHub.
If not, then specify it here.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Add a user via their GitHub username
}
