% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{add_secret}
\alias{add_secret}
\title{Add a new secret to the vault}
\usage{
add_secret(name, value, users, vault = NULL)
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{value}{Value of the secret, an arbitrary R object that
will be serialized using \code{\link[base:serialize]{base::serialize()}}.}

\item{users}{Email addresses of users that will have access to the
secret. (See \code{\link[=add_user]{add_user()}})}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
By default, the newly added secret is not shared with other
users. See the users argument if you want to change this.
You can also use \code{\link[=share_secret]{share_secret()}} later, to specify the users that
have access to the secret.
}
\seealso{
Other secret functions: \code{\link{delete_secret}},
  \code{\link{get_secret}}, \code{\link{list_secrets}},
  \code{\link{local_key}}, \code{\link{share_secret}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
