% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{add_secret}
\alias{add_secret}
\title{Add a new secret to the vault}
\usage{
add_secret(name, value, email, vault = NULL)
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{value}{Value of the secret, an arbitrary R object that
will be saved to a file using \code{\link[base:save]{base::save()}}.}

\item{email}{User that will have access to the secret.}

\item{vault}{Vault location. TODO}
}
\description{
By default, the newly added secret is not shared with other
users. See the users argument if you want to change this.
You can also use \code{\link[=share_secret]{share_secret()}} later, to specify the users that
have access to the secret.
}
