% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_travis_user}
\alias{add_travis_user}
\title{Add a user via their Travis repo}
\usage{
add_travis_user(travis_repo, email = travis_id, vault = NULL)
}
\arguments{
\item{travis_repo}{Name of Travis repository, usually in a format
<<username>>/<<repo>>}

\item{email}{Email address.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Add a user via their Travis repo
}
\seealso{
Other user functions: \code{\link{add_github_user}},
  \code{\link{add_user}}, \code{\link{delete_user}},
  \code{\link{list_users}}
}
