% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user}
\alias{add_user}
\title{Add a new user to the vault}
\usage{
add_user(email, public_key, vault = NULL)
}
\arguments{
\item{email}{Email address of the user. This is used to identify
users.}

\item{public_key}{Public key of the user. This is used to encrypt
the secrets for the different users. It can be
\itemize{
\item a string containing a PEM,
\item a file name that points to a PEM file,
\item a \code{pubkey} object created via the \code{openssl} package.
}}

\item{vault}{Vault location. TODO}
}
\description{
By default the new user does not have access to any secrets.
See \code{\link[=share_secret]{share_secret()}} to give them access.
}
