% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{delete_secret}
\alias{delete_secret}
\title{Remove a secret from the vault.}
\usage{
delete_secret(name, vault = NULL)
}
\arguments{
\item{name}{Name of the secret to delete.}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
Remove a secret from the vault.
}
\seealso{
Other secret functions: \code{\link{add_secret}},
  \code{\link{get_secret}}, \code{\link{list_secrets}},
  \code{\link{local_key}}, \code{\link{share_secret}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
