% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user}
\usage{
delete_user(email, vault = NULL)
}
\arguments{
\item{email}{Email address of the user.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
It also removes access of the user to all secrets, so if the user
is re-added again, they will not have access to any secrets.
}
