% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{get_secret}
\alias{get_secret}
\title{Retrieve a secret from the vault}
\usage{
get_secret(name, key = local_key(), vault = NULL)
}
\arguments{
\item{name}{Name of the secret.}

\item{key}{The private RSA key to use. It defaults to the current
user's default key.}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
Retrieve a secret from the vault
}
\seealso{
Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{list_secrets}},
  \code{\link{local_key}}, \code{\link{share_secret}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
