% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{get_secret}
\alias{get_secret}
\title{Retrieve a secret from the vault}
\usage{
get_secret(name, key = my_key(), vault = NULL)
}
\arguments{
\item{name}{Name of the secret.}

\item{key}{The private RSA key to use. It defaults to the current
user's default key.}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Retrieve a secret from the vault
}
