% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{list_secrets}
\alias{list_secrets}
\title{List all secrets}
\usage{
list_secrets(vault = NULL)
}
\arguments{
\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
List all secrets
}
