% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{list_secrets}
\alias{list_secrets}
\title{List all secrets}
\usage{
list_secrets(vault = NULL)
}
\arguments{
\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
List all secrets
}
\seealso{
Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{get_secret}},
  \code{\link{local_key}}, \code{\link{share_secret}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
