% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_users}
\alias{list_users}
\title{List users}
\usage{
list_users(vault = NULL)
}
\arguments{
\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
List users
}
