% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_users}
\alias{list_users}
\title{List users}
\usage{
list_users(vault = NULL)
}
\arguments{
\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
List users
}
\seealso{
Other user functions: \code{\link{add_github_user}},
  \code{\link{add_travis_user}}, \code{\link{add_user}},
  \code{\link{delete_user}}
}
