% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{share_secret}
\alias{share_secret}
\title{Share a secret among some users}
\usage{
share_secret(name, users, key = local_key(), vault = NULL)
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{users}{Character vector of usernames or email addresses
to share the secret with.}

\item{key}{Private key that has access to the secret. (I.e. its
corresponding public key is among the vault users.)}

\item{vault}{Vault location. Currently, vaults must be in R packages,
in the \code{inst/vault} directory.}
}
\description{
Use this function to extend the set of users that have access to a
secret. The calling user must have access to the secret as well.
}
\seealso{
\code{\link[=unshare_secret]{unshare_secret()}}

Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{get_secret}},
  \code{\link{list_secrets}}, \code{\link{local_key}},
  \code{\link{unshare_secret}}, \code{\link{update_secret}}
}
