% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{update_secret}
\alias{update_secret}
\title{Update a secret in the vault.}
\usage{
update_secret(name, value, key = local_key(), vault = NULL)
}
\arguments{
\item{name}{Name of the secret.}

\item{value}{Value of the secret, an arbitrary R object that
will be serialized using \code{\link[base:serialize]{base::serialize()}}.}

\item{key}{The private RSA key to use. It defaults to the current
user's default key.}

\item{vault}{Vault location. To create a vault, use \code{\link[=create_vault]{create_vault()}} or \code{\link[=create_package_vault]{create_package_vault()}}}
}
\description{
Update a secret in the vault.
}
\seealso{
Other secret functions: \code{\link{add_secret}},
  \code{\link{delete_secret}}, \code{\link{get_secret}},
  \code{\link{list_secrets}}, \code{\link{local_key}},
  \code{\link{share_secret}}, \code{\link{unshare_secret}}
}
