% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{update_secret}
\alias{update_secret}
\title{Update a secret in the vault.}
\usage{
update_secret(name, vault = NULL)
}
\arguments{
\item{name}{Name of the secret, a string that can contain alphanumeric
characters, underscores, dashes and dots.}

\item{vault}{Vault location. TODO}
}
\description{
Update a secret in the vault.
}
